requirements.txt
spotdl
pyyaml
